# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Sistem Akademik Mahasiswa API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Mahasiswa {
  id: number;
  nim: string;
  nama: string;
  jurusan: string;
  angkatan: number;
  /** @nullable */
  email?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface MahasiswaInput {
  /** @minLength 1 */
  nim: string;
  /** @minLength 1 */
  nama: string;
  /** @minLength 1 */
  jurusan: string;
  angkatan: number;
  /**
     * @nullable
     * @pattern ^[^\s@]+@[^\s@]+\.[^\s@]+$
     */
  email?: string | null;
}

export interface MahasiswaUpdate {
  /** @minLength 1 */
  nama?: string;
  /** @minLength 1 */
  jurusan?: string;
  angkatan?: number;
  /**
     * @nullable
     * @pattern ^[^\s@]+@[^\s@]+\.[^\s@]+$
     */
  email?: string | null;
}

export interface MahasiswaListResponse {
  data: Mahasiswa[];
  total: number;
}

export interface MataKuliah {
  id: number;
  kodeMk: string;
  namaMk: string;
  sks: number;
  /** @nullable */
  dosenPengampu?: string | null;
  createdAt: string;
}

export interface MataKuliahInput {
  /** @minLength 1 */
  kodeMk: string;
  /** @minLength 1 */
  namaMk: string;
  /** @minimum 1 */
  sks: number;
  /** @nullable */
  dosenPengampu?: string | null;
}

export interface MataKuliahUpdate {
  /** @minLength 1 */
  namaMk?: string;
  /** @minimum 1 */
  sks?: number;
  /** @nullable */
  dosenPengampu?: string | null;
}

export interface MataKuliahListResponse {
  data: MataKuliah[];
  total: number;
}

export interface NilaiDetail {
  id: number;
  mahasiswaId: number;
  mataKuliahId: number;
  nilaiHuruf: string;
  semester: string;
  createdAt: string;
  updatedAt: string;
  mahasiswa: Mahasiswa;
  mataKuliah: MataKuliah;
}

export interface NilaiInput {
  mahasiswaId: number;
  mataKuliahId: number;
  /** @minLength 1 */
  nilaiHuruf: string;
  /** @minLength 1 */
  semester: string;
}

export interface NilaiUpdate {
  /** @minLength 1 */
  nilaiHuruf?: string;
  /** @minLength 1 */
  semester?: string;
}

export interface NilaiListResponse {
  data: NilaiDetail[];
  total: number;
}

export interface TranskripItem {
  nilaiId: number;
  mataKuliah: MataKuliah;
  nilaiHuruf: string;
  semester: string;
  bobotNilai: number;
}

export interface Transkrip {
  mahasiswa: Mahasiswa;
  nilai: TranskripItem[];
  totalSks: number;
  ipk: number;
}

export interface NilaiDistribusi {
  nilaiHuruf: string;
  jumlah: number;
}

export interface DashboardStats {
  totalMahasiswa: number;
  totalMataKuliah: number;
  totalNilai: number;
  distribusiNilai: NilaiDistribusi[];
  mahasiswaTerbaru: Mahasiswa[];
}

export interface Admin {
  id: string;
  email: string;
  nama: string;
}

export interface AdminLoginInput {
  /**
     * @nullable
     * @pattern ^[^\s@]+@[^\s@]+\.[^\s@]+$
     */
  email: string | null;
  /** @minLength 8 */
  password: string;
}

export interface AdminRegisterInput {
  /**
     * @nullable
     * @pattern ^[^\s@]+@[^\s@]+\.[^\s@]+$
     */
  email: string | null;
  /** @minLength 8 */
  password: string;
  /** @minLength 1 */
  nama: string;
}

export interface LogoutResponse {
  ok: boolean;
}

export type ListMahasiswaParams = {
limit?: number;
offset?: number;
search?: string;
};

export type ListMataKuliahParams = {
limit?: number;
offset?: number;
};

export type ListNilaiParams = {
mahasiswa_id?: number;
limit?: number;
offset?: number;
};