# lib/api-zod/src/generated/types/admin.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Sistem Akademik Mahasiswa API
 * OpenAPI spec version: 0.1.0
 */

export interface Admin {
  id: string;
  email: string;
  nama: string;
}