# lib/api-zod/src/generated/types/adminLoginInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Sistem Akademik Mahasiswa API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminLoginInput {
  /**
     * @nullable
     * @pattern ^[^\s@]+@[^\s@]+\.[^\s@]+$
     */
  email: string | null;
  /** @minLength 8 */
  password: string;
}