# lib/api-zod/src/generated/types/mahasiswaInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Sistem Akademik Mahasiswa API
 * OpenAPI spec version: 0.1.0
 */

export interface MahasiswaInput {
  /** @minLength 1 */
  nim: string;
  /** @minLength 1 */
  nama: string;
  /** @minLength 1 */
  jurusan: string;
  angkatan: number;
  /**
     * @nullable
     * @pattern ^[^\s@]+@[^\s@]+\.[^\s@]+$
     */
  email?: string | null;
}