# lib/api-zod/src/generated/types/mataKuliahInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Sistem Akademik Mahasiswa API
 * OpenAPI spec version: 0.1.0
 */

export interface MataKuliahInput {
  /** @minLength 1 */
  kodeMk: string;
  /** @minLength 1 */
  namaMk: string;
  /** @minimum 1 */
  sks: number;
  /** @nullable */
  dosenPengampu?: string | null;
}