# lib/api-zod/src/generated/types/transkrip.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Sistem Akademik Mahasiswa API
 * OpenAPI spec version: 0.1.0
 */
import type { Mahasiswa } from './mahasiswa';
import type { TranskripItem } from './transkripItem';

export interface Transkrip {
  mahasiswa: Mahasiswa;
  nilai: TranskripItem[];
  totalSks: number;
  ipk: number;
}